requirements.txt
amqp==5.0.6
asgiref==3.3.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.42.0
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-jalali==4.0.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-redis==4.12.1
django-registration==3.1.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-taggit==1.3.0
django-webpack-loader==0.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
itypes==1.2.0
jdatetime==3.6.2
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==8.0.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
